# feat(repository): improved zstd decompression memory usage