```
$ kopia benchmark compression --data-file ~/Downloads/kopia-index-inspect-all.log --algorithms zstd,zstd-fastest,zstd-better-compression,zstd-best-compression --parallel 16 --operations=decompress --repeat=10
```

BEFORE:

```
     Compression                Compressed   Throughput   Allocs   Memory Usage
------------------------------------------------------------------------------------------------
  0. zstd-fastest               18.4 MB      6.2 GB/s     22539    5.5 GB
  1. zstd-better-compression    16.4 MB      5.1 GB/s     18252    6.3 GB
  2. zstd                       18.1 MB      4.7 GB/s     17933    6.4 GB
  3. zstd-best-compression      16.2 MB      4 GB/s       18632    6.3 GB (deprecated)
```

AFTER

```
     Compression                Compressed   Throughput   Allocs   Memory Usage
------------------------------------------------------------------------------------------------
  0. zstd-fastest               18.4 MB      7.8 GB/s     1078     4.3 GB
  1. zstd-best-compression      16.2 MB      5.8 GB/s     1265     4.4 GB (deprecated)
  2. zstd-better-compression    16.4 MB      5.7 GB/s     1250     4.4 GB
  3. zstd                       18.1 MB      5.4 GB/s     1570     4.5 GB
```