@jkowalski 2024-04-05
superseded by https://github.com/kopia/kopia/pull/3780